Phase 3: Extract model layer from utils/metadata.py, break last circular import
- Move ModMetadata, ModReplacement, WorkshopUpdateResult data types to
  models/metadata/metadata_structure.py
- Move recursively_update_dict to new utils/dict_utils.py
- Hoist DBBuilderCore import in metadata.py (no more cycle)
- Hoist recursively_update_dict import in db_builder_core.py (no more cycle)
- Create models/metadata/__init__.py with backward-compat re-exports
- Remove stale metadata ↔ db_builder_core entries from checker ALLOWED set
- Metadata in utils/metadata.py shrunk from 2974 to 2887 lines
- Deferred imports in both files: 2 → 0
