+++ app/models/metadata/__init__.py
@@ -1,0 +1,43 @@
+from app.models.metadata.metadata_structure import (
+    AboutXmlMod,
+    BaseRules,
+    CaseInsensitiveSet,
+    CaseInsensitiveStr,
+    DependencyMod,
+    ExternalRule,
+    ExternalRulesSchema,
+    ListedMod,
+    ModMetadata,
+    ModReplacement,
+    ModsConfig,
+    ModType,
+    Rules,
+    ScenarioMod,
+    SteamDbEntry,
+    SteamDbEntryBlacklist,
+    SteamDbEntryDependency,
+    SteamDbSchema,
+    WorkshopUpdateResult,
+)
+
+__all__ = [
+    "AboutXmlMod",
+    "BaseRules",
+    "CaseInsensitiveSet",
+    "CaseInsensitiveStr",
+    "DependencyMod",
+    "ExternalRule",
+    "ExternalRulesSchema",
+    "ListedMod",
+    "ModMetadata",
+    "ModReplacement",
+    "ModsConfig",
+    "ModType",
+    "Rules",
+    "ScenarioMod",
+    "SteamDbEntry",
+    "SteamDbEntryBlacklist",
+    "SteamDbEntryDependency",
+    "SteamDbSchema",
+    "WorkshopUpdateResult",
+]
