+++ app/utils/dict_utils.py
@@ -1,0 +1,50 @@
+from typing import Any, Iterable
+
+
+def recursively_update_dict(
+    a_dict: dict[str, Any],
+    b_dict: dict[str, Any],
+    prune_exceptions: Iterable[str] = [],
+    purge_keys: Iterable[str] = [],
+    recurse_exceptions: Iterable[str] = [],
+) -> None:
+    # Check for keys in recurse_exceptions in a_dict that are not in b_dict and remove them
+    for key in set(a_dict.keys()) - set(b_dict.keys()):
+        if recurse_exceptions and key in recurse_exceptions:
+            del a_dict[key]
+    # Recursively update A with B, excluding recurse exceptions (list of keys to just overwrite)
+    for key, value in b_dict.items():
+        if recurse_exceptions and key in recurse_exceptions:
+            # If the key is an exception, update its value directly from B
+            a_dict[key] = value
+        elif (
+            key in a_dict and isinstance(a_dict[key], dict) and isinstance(value, dict)
+        ):
+            # If the key exists in both dictionaries and the values are dictionaries,
+            # recursively update the nested dictionaries except for the recurse exceptions list
+            recursively_update_dict(
+                a_dict[key],
+                value,
+                prune_exceptions=prune_exceptions,
+                purge_keys=purge_keys,
+                recurse_exceptions=recurse_exceptions,
+            )
+        else:
+            # Otherwise, update the value in A with the value from B
+            a_dict[key] = value
+    # Prune keys with empty dictionary values (except for keys in prune exceptions list)
+    keys_to_delete = [
+        key
+        for key, value in a_dict.items()
+        if isinstance(value, dict)
+        and not value
+        and (prune_exceptions is None or key not in prune_exceptions)
+    ]
+    for key in keys_to_delete:
+        del a_dict[key]
+    # Delete keys from the list of keys to delete
+    if purge_keys is not None:
+        for key in purge_keys:
+            if key in a_dict:
+                del a_dict[key]
+    return None
